fix: その他のページも詳細表示のレイアウトを調整
